Refactoring visitForm
