+++ frontend/src/App.js
@@ -10,7 +10,7 @@ import UserForm from './components/UserForm'
 import { CURRENT_USER } from './graphql/queries'
 import UserList from './components/UserList'
 import EventForm from './components/EventForm'
-import VisitForm from './components/VisitForm'
+import { VisitForm } from './components/VisitForm'
 import { FcKey } from 'react-icons/fc'
 import UserPage from './components/UserPage'
 import EventPage from './components/EventPage'
